WIP Add undo
We add two "undo_list_t", one for undo and one for redo.

If we do anything that changes the commandline, we add an undo
element, if we undo, we add a redo element.

Also we coalesce small changes like normal string insertions and
single-character deletions.

This is only lightly tested and might not work for everything.

Fixes #1367.