Avoid recursion in complete -C given an empty argument
In completions that do not use the current commandline:
when given $arg that might be empty, we need to call

`complete -C "$arg"`

If "$arg" is an empty string or has zero elements, both `complete
-C"$arg"` and `complete -C $arg` can be equivalent to `complete -C`
which uses the current commandline.  In this case, complete might call
itself recursively, as can be observed by typing `sudo <TAB>`.