Add completion_result_t specifying the completed token
Unlike in early versions of fish, we now have many different kinds of
special symbols available that act as redirections, pipes, and
statement separators. Since most of them contain a special symbols
like &, | and >, we can make them discoverable by adding completions
that trigger if the cursor is in a string that contains such symbols
(and that is not recognised as normal token).

Up to now, <&|> are not handled in completions at all.  These symbols
are not printed by `builtin commandline -o` and not recognized as
tokens by parse_util_token_extent(). The latter is used to determine
the token to complete (and possibly replace), so some changes are
needed to complete special symbols.

The function complete() knows where on the commandline to place a
completion, but this information was not returned to the caller, who
had to call parse_util_token_extent again to find out where to insert
the completion.  When completing things like 2>&1, it will be even more
complicated to compute the place of insertions because they are not
handled by the tokenize.  So we just return it in complete(). The info
is just a pair of indices that specify the range of the token that can
be replaced by a completion.

So we need to share the source range of the token to complete.
While we're at it we can get rid of some redundant calls to
parse_util_cmdsubst_extent() that were done before calling complete(),
which does it anyway.

This also changes completion_apply_to_command_line to take a range
instead of a simple cursor position; if the completion is a replacement,
it will replace that range instead of computing the token extent from
the cursor position.