Migrate some iothread functions into member functions of thread_pool_t
This reintroduces commits 22230a1a0ddddc62732418773f3741073988df96
and 9d7d70c204073fb7fcfc393b8b78821b905e6817, now with the bug fixed.

The problem was when there was one thread waiting in the pool. We enqueue
an item onto the pool and attempt to wake up the thread. But before the
thread runs, we enqueue another item - this second enqueue will see the
thread waiting and attempt to wake it up as well. If the two work items
were dependent (reader/writer) then we would have a deadlock.

The fix is to check if the number of waiting threads is at least as large
as the queue. If the number of enqueued items exceeds the number of waiting
threads, then spawn a new thread always.