Add the machine "type" to the default prompt
This is a description of whether we're connected over ssh or in a
container or something.

It's used to signal to the user that this
particular terminal is "special", and not on the local machine, e.g.
to avoid mistakes like executing `poweroff` on a server instead of
your development laptop.

Unfortunately this (at least currently) isn't dependable enough
to *disable* display of the user@host part on "normal" machines, but
we can add something on ssh.

See #6375.