Make the history search term editable
Summary: after starting a history search by pressing the up arrow,
the highlighted search term can be edited. After doing so, pressing
up again will resume the search from where you left off but with the
updated search term.

When doing a history search with an empty search string, the behavior
remains unchanged.  In particular, the cursor is positioned at the end
of the line after moving to another history entry.

When doing history search with a non-empty search string, the cursor will
be positioned at the end of the matching search string.  Subsequently,
the search term can be edited as long as one only inserts or deletes
text inside the highlighted search string, or inserts text right after
the search string. Other edits end history search as usual.

The ability to edit the search term allows to clear it and then navigate
in the vicinity of the currently matched history entry.
For example, imagine you have a history like

	echo Some command I cannot recall
	echo Some command I do remember
	...

And you want to recall the first command, but remember neither its name
nor any of its arguments.  Now you can simply type "remember" and press
the up arrow to go to the second command.  Then delete the search string
by pressing Control-W and press the up arrow to go to the first command.

Note that edits to the search string are part of the search's transient
edit, which means they will all be undone with a single undo
(every search should create exactly one atomic undo group).