Incremental search
When pressing Control-R, fish will start a new incremental search,
(if a search is already active it will be turned into an incremental one).

Incremental search means that whenever you change the search term and it
does not match the current history entry anymore, then fish will
continue the search (in the same direction as the previous search command).

Closes #602