history search: do not move the cursor after failing backward search
When editing a multiline command line and pressing "up" too many times,
it can be somewhat annoying when the cursor is moved to the end because
fish attempted a history search and (likely) failed. Just don't move
the cursor (nor repaint) when the search fails.

This will also be important when we allow resuming search after editing
the search string. We don't want a failing backward search to replace
the commandline with just the search string. That should only happen
on failing forward search (or end-of-history) for consistency with
normal up-arrow search.